# backend/openapi/SalesDashboard.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SalesDashboard.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/dashboard/kpis": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Dashboard KPI cards for the selected period versus the previous one.",
        "operationId": "GetKpis",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KpiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/managers/ranking": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Managers ranked by gross profit or average check for the selected period.",
        "operationId": "GetManagerRanking",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "rankBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/timeseries": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Revenue, gross profit and Paid sales count over time, bucketed by a server-chosen step.",
        "operationId": "GetTimeseries",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeseriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/categories": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Revenue, profit, margin, units and revenue share per product category.",
        "operationId": "GetCategories",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/products/top": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "The 10 products with the highest gross profit for the selected period.",
        "operationId": "GetTopProducts",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopProductsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/sales/recent": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Keyset-paginated feed of recent sales (all statuses) for the selected period.",
        "operationId": "GetRecentSales",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentSalesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AverageCheckMetric": {
        "required": [
          "current",
          "previous",
          "change"
        ],
        "type": "object",
        "properties": {
          "current": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "previous": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "change": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "BestManagerDto": {
        "required": [
          "managerId",
          "firstName",
          "lastName",
          "grossProfit",
          "tiedCount"
        ],
        "type": "object",
        "properties": {
          "managerId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "grossProfit": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "tiedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CategoriesResponse": {
        "required": [
          "period",
          "items"
        ],
        "type": "object",
        "properties": {
          "period": {
            "$ref": "#/components/schemas/PeriodDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryRow"
            }
          }
        }
      },
      "CategoryRow": {
        "required": [
          "categoryId",
          "name",
          "revenue",
          "grossProfit",
          "margin",
          "unitsSold",
          "revenueShare"
        ],
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "revenue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "grossProfit": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "margin": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "unitsSold": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "revenueShare": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "CountMetric": {
        "required": [
          "current",
          "previous",
          "change"
        ],
        "type": "object",
        "properties": {
          "current": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "previous": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "change": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "KpiResponse": {
        "required": [
          "period",
          "previousPeriod",
          "revenue",
          "grossProfit",
          "salesCount",
          "averageCheck",
          "margin",
          "refunds",
          "bestManager"
        ],
        "type": "object",
        "properties": {
          "period": {
            "$ref": "#/components/schemas/PeriodDto"
          },
          "previousPeriod": {
            "$ref": "#/components/schemas/PeriodDto"
          },
          "revenue": {
            "$ref": "#/components/schemas/ValueMetric"
          },
          "grossProfit": {
            "$ref": "#/components/schemas/ValueMetric"
          },
          "salesCount": {
            "$ref": "#/components/schemas/CountMetric"
          },
          "averageCheck": {
            "$ref": "#/components/schemas/AverageCheckMetric"
          },
          "margin": {
            "$ref": "#/components/schemas/MarginMetric"
          },
          "refunds": {
            "$ref": "#/components/schemas/RefundsDto"
          },
          "bestManager": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BestManagerDto"
              }
            ]
          }
        }
      },
      "MarginMetric": {
        "required": [
          "current",
          "previous",
          "changePoints"
        ],
        "type": "object",
        "properties": {
          "current": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "previous": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "changePoints": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "PeriodDto": {
        "required": [
          "from",
          "to"
        ],
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RankingResponse": {
        "required": [
          "rankBy",
          "period",
          "previousPeriod",
          "items"
        ],
        "type": "object",
        "properties": {
          "rankBy": {
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/PeriodDto"
          },
          "previousPeriod": {
            "$ref": "#/components/schemas/PeriodDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RankingRow"
            }
          }
        }
      },
      "RankingRow": {
        "required": [
          "rank",
          "managerId",
          "firstName",
          "lastName",
          "team",
          "isActive",
          "hasSales",
          "salesCount",
          "revenue",
          "grossProfit",
          "averageCheck",
          "margin",
          "metricChange"
        ],
        "type": "object",
        "properties": {
          "rank": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "managerId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "team": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "hasSales": {
            "type": "boolean"
          },
          "salesCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "revenue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "grossProfit": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageCheck": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "margin": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "metricChange": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "RecentSaleItemDto": {
        "required": [
          "productName",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "productName": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "RecentSaleRow": {
        "required": [
          "saleId",
          "soldAt",
          "managerId",
          "managerName",
          "customerCompany",
          "status",
          "amount",
          "grossProfit",
          "items"
        ],
        "type": "object",
        "properties": {
          "saleId": {
            "type": "string",
            "format": "uuid"
          },
          "soldAt": {
            "type": "string",
            "format": "date-time"
          },
          "managerId": {
            "type": "string",
            "format": "uuid"
          },
          "managerName": {
            "type": "string"
          },
          "customerCompany": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SaleStatus"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "grossProfit": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentSaleItemDto"
            }
          }
        }
      },
      "RecentSalesResponse": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentSaleRow"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RefundsDto": {
        "required": [
          "amount",
          "rate"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/ValueMetric"
          },
          "rate": {
            "$ref": "#/components/schemas/MarginMetric"
          }
        }
      },
      "SaleStatus": {
        "type": "integer"
      },
      "TimeseriesPoint": {
        "required": [
          "bucketStart",
          "bucketEnd",
          "revenue",
          "grossProfit",
          "salesCount"
        ],
        "type": "object",
        "properties": {
          "bucketStart": {
            "type": "string",
            "format": "date"
          },
          "bucketEnd": {
            "type": "string",
            "format": "date"
          },
          "revenue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "grossProfit": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "salesCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "TimeseriesResponse": {
        "required": [
          "granularity",
          "period",
          "points"
        ],
        "type": "object",
        "properties": {
          "granularity": {
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/PeriodDto"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeseriesPoint"
            }
          }
        }
      },
      "TopProductRow": {
        "required": [
          "productId",
          "sku",
          "name",
          "categoryName",
          "revenue",
          "grossProfit",
          "margin",
          "unitsSold"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "revenue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "grossProfit": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "margin": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "unitsSold": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "TopProductsResponse": {
        "required": [
          "period",
          "items"
        ],
        "type": "object",
        "properties": {
          "period": {
            "$ref": "#/components/schemas/PeriodDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopProductRow"
            }
          }
        }
      },
      "ValueMetric": {
        "required": [
          "current",
          "previous",
          "change"
        ],
        "type": "object",
        "properties": {
          "current": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "previous": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "change": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Dashboard"
    }
  ]
}